Update after review comments
