+++ saleor/graphql/core/types/common.py
@@ -14,6 +14,7 @@ from ..enums import (
     ProductErrorCode,
     ShippingErrorCode,
     ShopErrorCode,
+    WebhookErrorCode,
 )
 from .money import VAT
 
@@ -75,6 +76,10 @@ class GiftCardError(Error):
 
 class ExtensionsError(Error):
     code = ExtensionsErrorCode(description="The error code.")
+
+
+class WebhookError(Error):
+    code = WebhookErrorCode(description="The error code.")
 
 
 class LanguageDisplay(graphene.ObjectType):
