+++ saleor/order/actions.py
@@ -8,12 +8,19 @@ from django.utils.translation import pgettext_lazy
 from ..core import analytics
 from ..extensions.manager import get_extensions_manager
 from ..payment import ChargeStatus, CustomPaymentChoices, PaymentError
-from . import FulfillmentStatus, OrderStatus, events, utils
+from ..product.utils import decrease_stock
+from . import FulfillmentStatus, OrderStatus, emails, events, utils
 from .emails import send_fulfillment_confirmation_to_customer, send_payment_confirmation
-from .utils import recalculate_order, restock_fulfillment_lines, update_order_status
+from .models import Fulfillment, FulfillmentLine
+from .utils import (
+    order_line_needs_automatic_fulfillment,
+    recalculate_order,
+    restock_fulfillment_lines,
+    update_order_status,
+)
 
 if TYPE_CHECKING:
-    from .models import Order, Fulfillment
+    from .models import Order
     from ..account.models import User, Address
     from ..payment.models import Payment
 
@@ -37,7 +44,7 @@ def handle_fully_paid_order(order: "Order"):
         send_payment_confirmation.delay(order.pk)
 
         if utils.order_needs_automatic_fullfilment(order):
-            utils.automatically_fulfill_digital_lines(order)
+            automatically_fulfill_digital_lines(order)
     try:
         analytics.report_order(order.tracking_client_id, order)
     except Exception:
@@ -207,3 +214,40 @@ def clean_mark_order_as_paid(order: "Order"):
                 "Orders with payments can not be manually marked as paid.",
             )
         )
+
+
+def fulfill_order_line(order_line, quantity):
+    """Fulfill order line with given quantity."""
+    if order_line.variant and order_line.variant.track_inventory:
+        decrease_stock(order_line.variant, quantity)
+    order_line.quantity_fulfilled += quantity
+    order_line.save(update_fields=["quantity_fulfilled"])
+
+
+def automatically_fulfill_digital_lines(order: "Order"):
+    """Fulfill all digital lines which have enabled automatic fulfillment setting.
+
+    Send confirmation email afterward.
+    """
+    digital_lines = order.lines.filter(
+        is_shipping_required=False, variant__digital_content__isnull=False
+    )
+    digital_lines = digital_lines.prefetch_related("variant__digital_content")
+
+    if not digital_lines:
+        return
+    fulfillment, _ = Fulfillment.objects.get_or_create(order=order)
+    for line in digital_lines:
+        if not order_line_needs_automatic_fulfillment(line):
+            continue
+        digital_content = line.variant.digital_content
+        digital_content.urls.create(line=line)
+        quantity = line.quantity
+        FulfillmentLine.objects.create(
+            fulfillment=fulfillment, order_line=line, quantity=quantity
+        )
+        fulfill_order_line(order_line=line, quantity=quantity)
+    emails.send_fulfillment_confirmation_to_customer(
+        order, fulfillment, user=order.user
+    )
+    update_order_status(order)
