+++ saleor/order/utils.py
@@ -13,14 +13,9 @@ from ..core.weight import zero_weight
 from ..dashboard.order.utils import get_voucher_discount_for_order
 from ..discount.models import NotApplicable
 from ..extensions.manager import get_extensions_manager
-from ..order import OrderStatus, emails
-from ..order.models import Fulfillment, FulfillmentLine, Order, OrderLine
-from ..product.utils import (
-    allocate_stock,
-    deallocate_stock,
-    decrease_stock,
-    increase_stock,
-)
+from ..order import OrderStatus
+from ..order.models import Order, OrderLine
+from ..product.utils import allocate_stock, deallocate_stock, increase_stock
 from ..product.utils.digital_products import get_default_digital_content_settings
 from ..shipping.models import ShippingMethod
 from . import events
@@ -44,43 +39,6 @@ def order_needs_automatic_fullfilment(order: Order) -> bool:
         if order_line_needs_automatic_fulfillment(line):
             return True
     return False
-
-
-def fulfill_order_line(order_line, quantity):
-    """Fulfill order line with given quantity."""
-    if order_line.variant and order_line.variant.track_inventory:
-        decrease_stock(order_line.variant, quantity)
-    order_line.quantity_fulfilled += quantity
-    order_line.save(update_fields=["quantity_fulfilled"])
-
-
-def automatically_fulfill_digital_lines(order: Order):
-    """Fulfill all digital lines which have enabled automatic fulfillment setting.
-
-    Send confirmation email afterward.
-    """
-    digital_lines = order.lines.filter(
-        is_shipping_required=False, variant__digital_content__isnull=False
-    )
-    digital_lines = digital_lines.prefetch_related("variant__digital_content")
-
-    if not digital_lines:
-        return
-    fulfillment, _ = Fulfillment.objects.get_or_create(order=order)
-    for line in digital_lines:
-        if not order_line_needs_automatic_fulfillment(line):
-            continue
-        digital_content = line.variant.digital_content
-        digital_content.urls.create(line=line)
-        quantity = line.quantity
-        FulfillmentLine.objects.create(
-            fulfillment=fulfillment, order_line=line, quantity=quantity
-        )
-        fulfill_order_line(order_line=line, quantity=quantity)
-    emails.send_fulfillment_confirmation_to_customer(
-        order, fulfillment, user=order.user
-    )
-    update_order_status(order)
 
 
 def check_order_status(func):
