+++ tests/test_order.py
@@ -12,14 +12,12 @@ from saleor.core.exceptions import InsufficientStock
 from saleor.core.weight import zero_weight
 from saleor.discount.utils import validate_voucher_in_order
 from saleor.order import OrderStatus, events as order_events, models
-from saleor.order.models import Fulfillment, Order
+from saleor.order.models import Order
 from saleor.order.templatetags.order_lines import display_translated_order_line_name
 from saleor.order.utils import (
     add_variant_to_order,
-    automatically_fulfill_digital_lines,
     change_order_line_quantity,
     delete_order_line,
-    fulfill_order_line,
     recalculate_order,
     restock_fulfillment_lines,
     restock_order_lines,
@@ -28,8 +26,8 @@ from saleor.order.utils import (
 )
 from saleor.payment import ChargeStatus
 from saleor.payment.models import Payment
-from saleor.product.models import DigitalContent
-from tests.utils import create_image, get_redirect_location
+
+from .utils import get_redirect_location
 
 
 def test_total_setter():
@@ -608,76 +606,6 @@ def test_get_order_weight_non_existing_product(order_with_lines, product):
     new_weight = order.get_total_weight()
 
     assert old_weight == new_weight
-
-
-@patch("saleor.order.utils.emails.send_fulfillment_confirmation")
-@patch("saleor.order.utils.get_default_digital_content_settings")
-def test_fulfill_digital_lines(
-    mock_digital_settings, mock_email_fulfillment, order_with_lines, media_root
-):
-    mock_digital_settings.return_value = {"automatic_fulfillment": True}
-    line = order_with_lines.lines.all()[0]
-
-    image_file, image_name = create_image()
-    variant = line.variant
-    digital_content = DigitalContent.objects.create(
-        content_file=image_file, product_variant=variant, use_default_settings=True
-    )
-
-    line.variant.digital_content = digital_content
-    line.is_shipping_required = False
-    line.save()
-
-    order_with_lines.refresh_from_db()
-    automatically_fulfill_digital_lines(order_with_lines)
-    line.refresh_from_db()
-    fulfillment = Fulfillment.objects.get(order=order_with_lines)
-    fulfillment_lines = fulfillment.lines.all()
-
-    assert fulfillment_lines.count() == 1
-    assert line.digital_content_url
-    assert mock_email_fulfillment.delay.called
-
-
-def test_fulfill_order_line(order_with_lines):
-    order = order_with_lines
-    line = order.lines.first()
-    quantity_fulfilled_before = line.quantity_fulfilled
-    variant = line.variant
-    stock_quantity_after = variant.quantity - line.quantity
-
-    fulfill_order_line(line, line.quantity)
-
-    variant.refresh_from_db()
-    assert variant.quantity == stock_quantity_after
-    assert line.quantity_fulfilled == quantity_fulfilled_before + line.quantity
-
-
-def test_fulfill_order_line_with_variant_deleted(order_with_lines):
-    line = order_with_lines.lines.first()
-    line.variant.delete()
-
-    line.refresh_from_db()
-
-    fulfill_order_line(line, line.quantity)
-
-
-def test_fulfill_order_line_without_inventory_tracking(order_with_lines):
-    order = order_with_lines
-    line = order.lines.first()
-    quantity_fulfilled_before = line.quantity_fulfilled
-    variant = line.variant
-    variant.track_inventory = False
-    variant.save()
-
-    # stock should not change
-    stock_quantity_after = variant.quantity
-
-    fulfill_order_line(line, line.quantity)
-
-    variant.refresh_from_db()
-    assert variant.quantity == stock_quantity_after
-    assert line.quantity_fulfilled == quantity_fulfilled_before + line.quantity
 
 
 @patch("saleor.discount.utils.validate_voucher")
